Fixed function's division
I changed type INT to Float because some cases will problably have results with broken numbers

+++ FuntionsCalculator.c
@@ -12,7 +12,7 @@ int subtraction(int num1, int num2)
     return num1 - num2;
 }
 
-int division(int num1, int num2)
+float division(float num1, float num2)
 {
     return num1 / num2;
 }
