+++ FuntionsCalculator.h
@@ -1,4 +1,4 @@
 int sum(int num1, int num2);
 int subtraction(int num1, int num2);
-int division(int num1, int num2);
+float division(float num1, float num2);
 int multiplication(int num1, int num2);
